Addition of META-INF/services/javax.sound.sampled.spi.AudioFileWriter file (without this file, the service provider don't work), some warnin
Addition of META-INF/services/javax.sound.sampled.spi.AudioFileWriter
file (without this file, the service provider don't work), some warning
cleanups.

+++ src/javaFlacEncoder/AudioStreamEncoder.java
@@ -60,7 +60,7 @@ public class AudioStreamEncoder {
     int channels = format.getChannels();
     boolean bigEndian = format.isBigEndian();
     boolean isSigned = format.getEncoding() == AudioFormat.Encoding.PCM_SIGNED;
-    byte[] samplesIn = new byte[(int)maxRead];
+    byte[] samplesIn = new byte[maxRead];
     int samplesRead;
     int framesRead;
     int[] sampleData = new int[maxRead*channels/frameSize];
